requirements.txt
requests==2.32.3
sqlalchemy==2.0.35
python-telegram-bot==20.7